修复了一个 Yaml 转换时 double 类型无法转为 float 类型的异常

+++ src/main/java/org/wolflink/minecraft/plugin/siriuxa/invbackup/PlayerBackpack.java
@@ -89,7 +89,7 @@ public class PlayerBackpack implements ConfigurationSerializable {
         boots = (ItemStack) map.get("boots");
         offhand = (ItemStack) map.get("offhand");
         level = (int) map.get("level");
-        exp = (float) map.get("exp");
+        exp = (float)((double) map.get("exp"));
         items = (ItemStack[]) map.get("items");
     }
     @NotNull
